requirements.txt
flask
flask-sqlalchemy
flask-wtf
flask-login
flask-bcrypt
psutil
paramiko
pymysql
wtforms
gunicorn
python-dotenv
requests
pyyaml
passlib
cryptography
pillow
matplotlib
jinja2